applied pr requests, split media operations to another interface. Compressed media files.
