+++ pom.xml
@@ -170,6 +170,11 @@
 			<artifactId>hibernate-search-backend-lucene</artifactId>
 			<version>7.1.1.Final</version>
 		</dependency>
+		<dependency>
+			<groupId>io.github.techgnious</groupId>
+			<artifactId>IVCompressor</artifactId>
+			<version>2.0.2</version>
+		</dependency>
 	</dependencies>
 	<build>
 		<plugins>
