+++ src/main/java/com/scalefocus/blogapplication/service/BlogService.java
@@ -2,12 +2,8 @@ package com.scalefocus.blogapplication.service;
 
 import com.scalefocus.blogapplication.dto.BlogPostDto;
 import com.scalefocus.blogapplication.dto.BlogPostSummaryDto;
-import com.scalefocus.blogapplication.dto.MediaFileDto;
 import com.scalefocus.blogapplication.dto.TagDto;
 import org.springframework.data.domain.Page;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
 
 public interface BlogService {
     BlogPostDto createBlog(BlogPostDto blogDto);
@@ -34,9 +30,4 @@ public interface BlogService {
 
     Page<BlogPostDto> searchBlogs(String query, int page, int size);
 
-    BlogPostDto addMediaFiles(Long id, List<MultipartFile> mediaFiles, List<MediaFileDto> mediaFileDtos);
-
-    BlogPostDto removeMediaFile(Long id, Long mediaFileId);
-
-    MediaFileDto getMediaFile(Long id, Long mediaFileId);
 }
