+++ src/test/java/com/scalefocus/blogapplication/service/integration/BlogServiceIntegrationTest.java
@@ -1,7 +1,6 @@
 package com.scalefocus.blogapplication.service.integration;
 
 import com.scalefocus.blogapplication.dto.*;
-import com.scalefocus.blogapplication.model.MediaType;
 import com.scalefocus.blogapplication.repository.BlogPostRepository;
 import com.scalefocus.blogapplication.repository.UserRepository;
 import com.scalefocus.blogapplication.service.BlogService;
@@ -13,19 +12,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import org.springframework.data.domain.Page;
-import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
-
-import javax.imageio.ImageIO;
-import java.awt.*;
-import java.awt.image.BufferedImage;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.util.List;
 
 import static com.scalefocus.blogapplication.mapper.BlogPostMapper.SUMMARY_LENGTH;
 import static org.junit.jupiter.api.Assertions.*;
@@ -183,98 +173,5 @@ class BlogServiceIntegrationTest {
         assertFalse(result.isEmpty());
         assertEquals(newBlog.getTitle(), result.getContent().get(0).getTitle());
         assertEquals("Content for summarized blog.".substring(0, SUMMARY_LENGTH), result.getContent().get(0).getSummary());
-    }
-
-    @Test
-    void testAddMediaFiles() throws IOException {
-        // Create a blog post
-        BlogPostDto newBlog = new BlogPostDto();
-        newBlog.setTitle("Blog with Media " + System.currentTimeMillis());
-        newBlog.setContent("Content for blog with media.");
-        BlogPostDto createdBlog = blogService.createBlog(newBlog);
-
-        BufferedImage image = new BufferedImage(100, 200, BufferedImage.TYPE_INT_RGB);
-        Graphics2D graphics = image.createGraphics();
-        graphics.setPaint(Color.BLUE);
-        graphics.fillRect(0, 0, image.getWidth(), image.getHeight());
-        graphics.dispose();
-
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        ImageIO.write(image, "jpg", baos);
-        baos.flush();
-        byte[] imageBytes = baos.toByteArray();
-        baos.close();
-
-        // Add media files to the blog post
-        MultipartFile multipartFile = new MockMultipartFile(
-                "file",
-                "image.jpg",
-                "image/jpeg",
-                imageBytes
-        );
-
-        MediaFileDto mediaFileDto = MediaFileDto.builder()
-                .content(imageBytes)
-                .mediaType(MediaType.IMAGE)
-                .width(100)
-                .height(200)
-                .size(multipartFile.getSize())
-                .build();
-
-        BlogPostDto updatedBlog = blogService.addMediaFiles(createdBlog.getId(), List.of(multipartFile), List.of(mediaFileDto));
-
-        assertNotNull(updatedBlog);
-        assertEquals(1, updatedBlog.getMediaFiles().size());
-        MediaFileDto addedMediaFile = updatedBlog.getMediaFiles().get(0);
-        assertEquals(MediaType.IMAGE, addedMediaFile.getMediaType());
-        assertEquals(multipartFile.getSize(), addedMediaFile.getSize());
-        assertEquals(100, addedMediaFile.getWidth());
-        assertEquals(200, addedMediaFile.getHeight());
-    }
-
-    @Test
-    void testRemoveMediaFile() throws IOException {
-        // Create a blog post
-        BlogPostDto newBlog = new BlogPostDto();
-        newBlog.setTitle("Blog with Media " + System.currentTimeMillis());
-        newBlog.setContent("Content for blog with media.");
-        BlogPostDto createdBlog = blogService.createBlog(newBlog);
-
-        BufferedImage image = new BufferedImage(100, 200, BufferedImage.TYPE_INT_RGB);
-        Graphics2D graphics = image.createGraphics();
-        graphics.setPaint(Color.BLUE);
-        graphics.fillRect(0, 0, image.getWidth(), image.getHeight());
-        graphics.dispose();
-
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        ImageIO.write(image, "jpg", baos);
-        baos.flush();
-        byte[] imageBytes = baos.toByteArray();
-        baos.close();
-
-        // Add media files to the blog post
-        MultipartFile multipartFile = new MockMultipartFile(
-                "file",
-                "image.jpg",
-                "image/jpeg",
-                imageBytes
-        );
-
-        MediaFileDto mediaFileDto = MediaFileDto.builder()
-                .content(imageBytes)
-                .mediaType(MediaType.IMAGE)
-                .width(100)
-                .height(200)
-                .size(multipartFile.getSize())
-                .build();
-
-        BlogPostDto updatedBlog = blogService.addMediaFiles(createdBlog.getId(), List.of(multipartFile), List.of(mediaFileDto));
-
-        // Remove the media file
-        Long mediaFileId = updatedBlog.getMediaFiles().get(0).getId();
-        BlogPostDto blogAfterRemoval = blogService.removeMediaFile(createdBlog.getId(), mediaFileId);
-
-        assertNotNull(blogAfterRemoval);
-        assertTrue(blogAfterRemoval.getMediaFiles().isEmpty());
     }
 }
